nlp-service/requirements.txt
# Kompl v2 nlp-service — commit 4. Pins per docs/research/2026-04-08-conversion-deps.md
# and docs/research/2026-04-09-llm-compile.md.
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
markitdown[all]==0.1.5
# Commit 4: LLM compile. Use google-genai NOT google-generativeai (deprecated 2025-Q1).
google-genai==1.12.1
# Commit 4: async token-bucket rate limiter. Single uvicorn worker only —
# InMemoryBucket is process-local; see research artifact section 3.
pyrate-limiter==3.7.0
